+++ packages/core/postgrest-js/test/types.generated.ts
@@ -308,7 +308,7 @@ export type Database = {
           created_at?: string
           data?: Json | null
           event_type?: string | null
-          id: number
+          id?: number
         }
         Update: {
           created_at?: string
@@ -329,7 +329,7 @@ export type Database = {
           created_at?: string
           data?: Json | null
           event_type?: string | null
-          id: number
+          id?: number
         }
         Update: {
           created_at?: string
@@ -786,22 +786,6 @@ export type Database = {
       }
       get_messages:
         | {
-            Args: { user_row: Database['public']['Tables']['users']['Row'] }
-            Returns: {
-              channel_id: number
-              data: Json | null
-              id: number
-              message: string | null
-              username: string
-            }[]
-            SetofOptions: {
-              from: 'users'
-              to: 'messages'
-              isOneToOne: false
-              isSetofReturn: true
-            }
-          }
-        | {
             Args: {
               channel_row: Database['public']['Tables']['channels']['Row']
             }
@@ -814,6 +798,22 @@ export type Database = {
             }[]
             SetofOptions: {
               from: 'channels'
+              to: 'messages'
+              isOneToOne: false
+              isSetofReturn: true
+            }
+          }
+        | {
+            Args: { user_row: Database['public']['Tables']['users']['Row'] }
+            Returns: {
+              channel_id: number
+              data: Json | null
+              id: number
+              message: string | null
+              username: string
+            }[]
+            SetofOptions: {
+              from: 'users'
               to: 'messages'
               isOneToOne: false
               isSetofReturn: true
@@ -917,22 +917,6 @@ export type Database = {
       }
       get_user_recent_messages:
         | {
-            Args: { user_row: Database['public']['Tables']['users']['Row'] }
-            Returns: {
-              channel_id: number | null
-              data: Json | null
-              id: number | null
-              message: string | null
-              username: string | null
-            }[]
-            SetofOptions: {
-              from: 'users'
-              to: 'recent_messages'
-              isOneToOne: false
-              isSetofReturn: true
-            }
-          }
-        | {
             Args: {
               active_user_row: Database['public']['Views']['active_users']['Row']
             }
@@ -945,6 +929,22 @@ export type Database = {
             }[]
             SetofOptions: {
               from: 'active_users'
+              to: 'recent_messages'
+              isOneToOne: false
+              isSetofReturn: true
+            }
+          }
+        | {
+            Args: { user_row: Database['public']['Tables']['users']['Row'] }
+            Returns: {
+              channel_id: number | null
+              data: Json | null
+              id: number | null
+              message: string | null
+              username: string | null
+            }[]
+            SetofOptions: {
+              from: 'users'
               to: 'recent_messages'
               isOneToOne: false
               isSetofReturn: true
@@ -997,8 +997,25 @@ export type Database = {
         Returns: number
       }
       postgrest_resolvable_with_override_function:
+        | { Args: never; Returns: undefined }
         | { Args: { a: string }; Returns: number }
         | { Args: { b: number }; Returns: string }
+        | {
+            Args: { cid: number; search?: string }
+            Returns: {
+              channel_id: number
+              data: Json | null
+              id: number
+              message: string | null
+              username: string
+            }[]
+            SetofOptions: {
+              from: '*'
+              to: 'messages'
+              isOneToOne: false
+              isSetofReturn: true
+            }
+          }
         | {
             Args: { profile_id: number }
             Returns: {
@@ -1028,37 +1045,20 @@ export type Database = {
               isSetofReturn: true
             }
           }
-        | {
-            Args: { cid: number; search?: string }
-            Returns: {
-              channel_id: number
-              data: Json | null
-              id: number
-              message: string | null
-              username: string
-            }[]
-            SetofOptions: {
-              from: '*'
-              to: 'messages'
-              isOneToOne: false
-              isSetofReturn: true
-            }
-          }
-        | { Args: never; Returns: undefined }
       postgrest_unresolvable_function:
-        | {
-            Args: { a: string }
-            Returns: {
-              error: true
-            } & 'Could not choose the best candidate function between: public.postgrest_unresolvable_function(a => int4), public.postgrest_unresolvable_function(a => text). Try renaming the parameters or the function itself in the database so function overloading can be resolved'
-          }
+        | { Args: never; Returns: undefined }
         | {
             Args: { a: number }
             Returns: {
               error: true
             } & 'Could not choose the best candidate function between: public.postgrest_unresolvable_function(a => int4), public.postgrest_unresolvable_function(a => text). Try renaming the parameters or the function itself in the database so function overloading can be resolved'
           }
-        | { Args: never; Returns: undefined }
+        | {
+            Args: { a: string }
+            Returns: {
+              error: true
+            } & 'Could not choose the best candidate function between: public.postgrest_unresolvable_function(a => int4), public.postgrest_unresolvable_function(a => text). Try renaming the parameters or the function itself in the database so function overloading can be resolved'
+          }
       set_users_offline: {
         Args: { name_param: string }
         Returns: {